Added getItemIntroductoryOfferDisplay function
