+++ client/me/purchases/manage-purchase/purchase-meta.jsx
@@ -302,7 +302,6 @@ function PurchaseMetaIntroductoryOfferDetail( { purchase } ) {
 	}
 
 	const text = getIntroductoryOfferIntervalDisplay(
-		translate,
 		purchase.introductoryOffer.intervalUnit,
 		purchase.introductoryOffer.intervalCount,
 		isIntroductoryOfferFreeTrial( purchase ),
